Create /templates folder
