+++ templates/meter/kostalpiko.yaml
